python-diarization/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pyannote.audio==3.1.1
torch==2.1.2
torchaudio==2.1.2
numpy==1.26.3
